requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Pillow
